requirements.txt
requests==2.31.0
colorama==0.4.6
rich==13.7.0
urllib3==2.0.7
nltk==3.8.1
textblob==0.17.1
wikipedia-api==0.6.0
duckduckgo-search==3.9.11 